backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.31.0
aiosqlite>=0.20.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic[email]>=2.10.3
alembic==1.14.0
httpx==0.28.1
pydantic-settings==2.6.1
fastapi-cache2==0.2.2
fake-useragent==1.5.1

tenacity==8.2.3
apscheduler==3.10.4
openai
tqdm>=4.66.0
slowapi>=0.1.9
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # pinned for passlib compatibility
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
aiosmtplib>=3.0.0
redis>=5.0.0
